refactor(auth): move PasswordResetTokenRepository interface to application/usecases
C4 из 6-step ADR-6 sequence. Interface relocates to consumer package
(same as C2/C3). password_reset_usecase.go uses bare name; `repositories`
import dropped (no longer needed in usecase file after C1 + C4).

Note: persistence/password_reset_token_repository_redis.go satisfies
the interface structurally — no import update needed (mock-pattern
file already free of `repositories` after C1 sentinel relocation).

Build / test / lint зелёные.
