+++ internal/modules/auth/application/usecases/password_reset_token_repository.go
@@ -1,4 +1,4 @@
-package repositories
+package usecases
 
 import (
 	"context"
